chore(form): Change file name to index.tsx
